heyvoca_back/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
werkzeug==2.0.3 
python-dotenv==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-Cors==4.0.1
SQLAlchemy==1.4.47
requests==2.28.2
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.129.0
gTTS==2.5.3
firebase-admin==6.5.0
pyfcm==2.0.7
APScheduler==3.10.4
pandas==1.5.3
PyMySQL==1.0.3